requirements.txt
scipy==1.13.1
PyYAML==6.0.1
Pillow==10.4.0
diffusers==0.18.2
transformers==4.44.1
peft==0.12.0
accelerate==1.3.0
safetensors==0.4.4
huggingface_hub==0.24.6
basicsr==1.4.2
timm==1.0.9
fairscale==0.4.13
loralib==0.1.2
clip==0.2.0
xformers==0.0.22.post7
